+++ tests/python_functional/src/driver_io/network/network_io.py
@@ -25,8 +25,8 @@ from enum import Enum
 from pathlib2 import Path
 
 import src.testcase_parameters.testcase_parameters as tc_parameters
+from src.common.file import File
 from src.common.random_id import get_unique_id
-from src.driver_io.file.file import File
 from src.helpers.loggen.loggen import Loggen
