+++ tests/python_functional/src/executors/command_executor.py
@@ -24,7 +24,7 @@ import logging
 
 import psutil
 
-from src.driver_io.file.file import File
+from src.common.file import File
 
 logger = logging.getLogger(__name__)
