+++ tests/python_functional/src/executors/process_executor.py
@@ -24,7 +24,7 @@ import logging
 
 import psutil
 
-from src.driver_io.file.file import File
+from src.common.file import File
 from src.executors.command_executor import prepare_executable_command
 from src.executors.command_executor import prepare_printable_command
 from src.executors.command_executor import prepare_std_outputs
